% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima_Graph.R
\name{arima_Graph}
\alias{arima_Graph}
\title{arima_DF}
\usage{
arima_Graph(DFI_completo, h)
}
\arguments{
\item{h}{numero de periodos a estimar}

\item{DF}{Data frame}
}
\description{
hacer un arima y pronosticar h periodos
Los datos ser?n agrupados a nivel horario
La salida es un DF con h periodos de mas, lo cual
}
