% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_miss_DF.R
\name{input_data_miss_DF}
\alias{input_data_miss_DF}
\title{input_data_miss_DF}
\usage{
input_data_miss_DF(DF)
}
\arguments{
\item{DF}{Data frame}
}
\value{
un data frame con datos imputados
}
\description{
sta funcion devuelve un DF con los datos imputados
}
