% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Time_dif.R
\name{Time_dif}
\alias{Time_dif}
\title{Time_dif}
\usage{
Time_dif(DF)
}
\arguments{
\item{DF}{Un DF cuya primera columna se llame FECHA y contenga datos posixct}
}
\value{
Una impresión con el tiempo transcurrido
}
\description{
Imprime el tiempo transcurrido desde que se encendio el sistema telemetrico
}
